Created chrome extension and added it to repo
-To use the extension, make sure to have developer mode on in the
chrome://extensions page.

-Click "Load Unpacked Extension" and navigate to the extension directory.

-Keep in mind that it isn't making use of actual classification results
thus far, but it does make use of my parser so that should give you and
idea of the delay to be expected.

-Furthermore, I will clean up some of the code when I get a chance to. I
focused heavily on making it possible to enable/disable the extension via
the click of a button. This allows users to turn it off for when they're
not performing medically related searches. It also allows them to clear
the current classifications made on the current google page.