Added finalized version of server and addon
-Server now preloads models instead of loading them on every
classification. This made things much faster especially when using the
true/false module.

-Added proper icon and removed redundant code in the addon. Changed the
name to BS Meter and it also now has a proper description.

-Added an additional piece of logic to the html_parser that fixes a glitch
with beautiful soup not being able to capture all of the relevant elements
that are requested. This helped reduce the number of errors shown by the
addon since some of the errors previously being thrown were due to there
not being any parsed text, even though the websites were full of text, so
the models wouldn't be able to perform classification.